[AGC] Recognize agc in triple parsing code
Added a new triple for the Apollo Guidance Computer (AGC) and allow the
triple parsing to recognise this triple and its corresponding
properties.

The target is not defined as little endian or big endian, since words
(16 bits) are not byte addressable, so byte ordering does not matter.
Assuming big endian to keep ordering within a word correct.

The target does have an assembler/linker already that uses a custom
object file format, however since we are providing the full toolchain,
using ELF is the most sensible choice.