[AGC] Support assembling section switching directives
This patch adds support for the BLOCK, BANK & BNKSUM directives
implemented as pseudo instructions. The BLOCK and BANK directives cause
an ELF section directive to be emitted in order to switch the current
output block or bank. BNKSUM is a directive intended to provide
information about the number of used words in a bank.