[AGC] Add support for materializing AGC constants
Since AGC can only materialize constants by inserting them in memory
using directives, it was required to override LLVM early on to ensure
that the necessary pattern can be emitted. The pattern uses a SETLOC to
change to the desired location, an OCT to output the raw constant and a
BANK to change back to the previous output address.