[AGC] Add basic CodeGen implementation
This patch adds the necessary files to implement a CodeGen target for
AGC. Only a very limited set of ALU operations are implemented so far.

The calling convention implemented here is not final. The goal is to
implement a virtual stack using one of the scratch registers and use
that to pass additional arguments.